feat: add CMakeLists.txt to build unit tests
In order for the unit tests to build on any platform there is now a
CMakeLists.txt made to build it.
In order for the tests to be able to run on any system they have been
updated to follow the naming conventions depending on the os.
The Github Actions have been updated to run on avery os and to push the
coverage only from ubuntu.
The .gitignore file has been updated to ignore build directories and
binaries.
The folder tests has been renamed to test in order to follow the
standard naming convention used by cmake test module detection.