chore: move `DYLIB_BUILD_TESTS` condition to make it easier for devs to contribute
`CMAKE_CURRENT_SOURCE_DIR STREQUAL CMAKE_SOURCE_DIR` restrains developers to use current directory as a cmake source.
This makes it harder for people with certain IDEs to work on the project.
By moving `DYLIB_BUILD_TESTS`, it allows other devs to be able to work on the project more consistently.